Repositorios y creacion de certificado

+++ src/main/java/com/vcore/backend_plataforma_web/repository/PrivilegioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.vcore.backend_plataforma_web.model.Privilegio;
 
-public interface PrivilegioRepository extends JpaRepository<Privilegio, Long>{
+public interface PrivilegioRepository extends JpaRepository<Privilegio, Integer>{
 
 }
